// web/src/api/model/deviceModelResponseMounting.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * nimap API
 * OpenAPI spec version: dev
 */

export type DeviceModelResponseMounting = typeof DeviceModelResponseMounting[keyof typeof DeviceModelResponseMounting];


export const DeviceModelResponseMounting = {
  rack: 'rack',
  wall: 'wall',
  desktop: 'desktop',
  din_rail: 'din_rail',
  ceiling: 'ceiling',
  pole: 'pole',
  embedded: 'embedded',
  other: 'other',
} as const;