// web/src/api/model/deviceModelResponsePoeInStandard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * nimap API
 * OpenAPI spec version: dev
 */

/**
 * The PoE standard the model needs when powered over Ethernet; only with poeIn
 */
export type DeviceModelResponsePoeInStandard = typeof DeviceModelResponsePoeInStandard[keyof typeof DeviceModelResponsePoeInStandard];


export const DeviceModelResponsePoeInStandard = {
  af: 'af',
  at: 'at',
  bt: 'bt',
  passive: 'passive',
} as const;